backend/requirements.txt
fastapi>=0.135.0
uvicorn[standard]
sqlalchemy[asyncio]>=2.0
asyncpg
alembic
python-jose[cryptography]
bcrypt
httpx
python-dotenv
pydantic[email]>=2.0
pydantic-settings>=2.0.0
python-multipart
redis>=5.0.0
